refactor: Change name 'Login' to 'Sign In'
