+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.login.SignInActivity"
+            android:name=".ui.signin.SignInActivity"
             android:exported="false" />
         <activity
             android:name=".ui.MainActivity"
